backend3/requirements.txt
fastapi==0.104.1
gql==3.4.1
requests==2.31.0
aiohttp==3.9.1
uvicorn==0.24.0.post1
neo4j==5.15.0
requests-toolbelt==1.0.0
llama-index==0.9.14
llama-hub==0.0.56
graphql-core==3.2.3
openai==1.3.9

